Add files via upload
A web-based tool called the Student Complaint 
Management System was created to make it easier for 
educational institutions to receive, monitor, and handle 
student complaints. Students can use the system to file 
complaints about a range of problems, including upkeep, 
technical assistance, housekeeping, etc. After that, these 
complaints are forwarded to the relevant departments for 
handling